Moved Signal definition to DigitalSchema
