+++ src/projects/digital/api/circuit/src/internal/assembly/components/OscilloscopeAssembler.ts
@@ -4,7 +4,7 @@ import {Schema} from "shared/api/circuit/schema";
 
 import {AssemblerParams, AssemblyReason} from "shared/api/circuit/internal/assembly/Assembler";
 
-import {Signal} from "digital/api/circuit/internal/sim/Signal";
+import {Signal} from "digital/api/circuit/schema/Signal";
 import {DigitalSim} from "digital/api/circuit/internal/sim/DigitalSim";
 
 import {DigitalComponentConfigurationInfo} from "../../DigitalComponents";
