+++ src/projects/digital/api/circuit/src/internal/sim/DigitalPropagators.ts
@@ -1,6 +1,8 @@
 import {Schema} from "shared/api/circuit/schema";
-import {Signal} from "./Signal";
-import {BCDtoDecimal, DecimalToBCD} from "../../utils/MathUtil";
+
+import {BCDtoDecimal, DecimalToBCD} from "digital/api/circuit/utils/MathUtil";
+import {Signal}                     from "digital/api/circuit/schema/Signal";
+
 import {PropagatorInfo, PropagatorsMap} from "./DigitalSim";
