+++ src/projects/digital/api/circuit/src/schema/DigitalSimState.ts
@@ -1,6 +1,5 @@
 import {GUID} from "shared/api/circuit/schema";
-// TODO[model_refactor_api] - define Signal in schema
-import {Signal} from "digital/api/circuit/internal/sim/Signal";
+import {Signal} from "./Signal";
 
 
 export interface DigitalSimState {
