+++ src/projects/digital/api/circuit/src/schema/export.ts
@@ -1,2 +1,3 @@
 export * from "./DigitalCircuit";
 export * from "./DigitalSimState";
+export * from "./Signal";
