+++ src/projects/digital/api/circuit/tests/sim/IC.test.ts
@@ -1,4 +1,4 @@
-import {Signal} from "digital/api/circuit/internal/sim/Signal";
+import {Signal} from "digital/api/circuit/schema/Signal";
 import "shared/tests/helpers/Extensions";
 
 import {CreateTestCircuit} from "tests/helpers/CreateTestCircuit";
